+++ components/boards/BoardList.tsx
@@ -2,7 +2,9 @@ import { formatDate } from "@/lib/formatDate";
 import React, { useEffect, useState } from "react";
 import styles from "./BoardList.module.css";
 import { Articles } from "@/lib/types";
-import Image from "next/image";
+
+import ArticleCard from "./ArticleCard";
+import BoardSearchSort from "./BoardSearchSort";
 
 interface BoardListProps {
   articles: Articles[];
@@ -12,8 +14,8 @@ interface BoardListProps {
 const BoardList = ({ articles, onOrderChange }: BoardListProps) => {
   const [searchQuery, setSearchQuery] = useState("");
 
-  const handelSearch = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setSearchQuery(e.target.value);
+  const handleSearch = (query: string) => {
+    setSearchQuery(query);
   };
 
   const filteredArticles = articles.filter((article) =>
@@ -26,74 +28,14 @@ const BoardList = ({ articles, onOrderChange }: BoardListProps) => {
         <p className={styles.title}>게시글</p>
         <button className={styles["write-button"]}>글쓰기</button>
       </div>
-      <div className={styles["search-box"]}>
-        <input
-          className={styles["search-bar"]}
-          placeholder="검색할 상품을 입력해주세요"
-          onChange={handelSearch}
-          value={searchQuery}
-        />
-        <div className={styles["select-wrapper"]}>
-          <select
-            className={styles["order-by-select"]}
-            onChange={(e) => onOrderChange(e.target.value)}
-          >
-            <option className={styles.option} value="recent">
-              최신순
-            </option>
-            <option className={styles.option} value="like">
-              좋아요순
-            </option>
-          </select>
-          <div className={styles["sort-icon"]}>
-            <Image
-              className={styles["image-component"]}
-              fill
-              src="/images/sortIcon.svg"
-              alt="화살표"
-            />
-          </div>
-        </div>
-      </div>
+      <BoardSearchSort
+        searchQuery={searchQuery}
+        onSearchChange={handleSearch}
+        onOrderChange={onOrderChange}
+      />
       <div className={styles["article-container"]}>
         {filteredArticles.map((article) => (
-          <div key={article.id} className={styles["article-box"]}>
-            <div className={styles["article-title-box"]}>
-              <p className={styles["article-title"]}>{article.title}</p>
-              <img
-                className={styles["product-img"]}
-                src={article.image}
-                alt="물품 이미지"
-              />
-            </div>
-            <div className={styles["info-box"]}>
-              <div className={styles["user-info-box"]}>
-                <div className={styles["user-profile"]}>
-                  <Image
-                    className={styles["image-component"]}
-                    fill
-                    src="/images/profileBig.svg"
-                    alt="프로필 이미지"
-                  />
-                </div>
-                <p className={styles["user-nickname"]}>
-                  {article.writer.nickname}
-                </p>
-                <p className={styles.date}>{formatDate(article.createdAt)}</p>
-              </div>
-              <div className={styles["like-count-box"]}>
-                <div className={styles.heart}>
-                  <Image
-                    className={styles["image-component"]}
-                    fill
-                    src="/images/heartIcon.svg"
-                    alt="좋아요 하트 이미지"
-                  />
-                </div>
-                <p className={styles["like-count"]}>{article.likeCount}</p>
-              </div>
-            </div>
-          </div>
+          <ArticleCard key={article.id} article={article} />
         ))}
       </div>
     </div>
