+++ components/boards/BoardSearchSort.module.css
@@ -1,0 +1,78 @@
+.search-box {
+  display: flex;
+  width: 100%;
+  align-items: center;
+  gap: 16px;
+  margin-bottom: 24px;
+
+  @media screen and (max-width: 744px) {
+    gap: 6px;
+    margin-bottom: 40px;
+    width: 100%;
+  }
+
+  @media screen and (max-width: 376px) {
+    gap: 13px;
+    margin-bottom: 16px;
+  }
+}
+
+.search-bar {
+  flex-grow: 1;
+  flex-shrink: 1;
+  height: 42px;
+  border: none;
+  border-radius: 12px;
+  padding: 9px 20px 9px 44px;
+  background-color: var(--color-white-400);
+  background-image: url("/images/searchIcon.png");
+  background-repeat: no-repeat;
+  background-position: left 16px center;
+  background-size: 20px;
+}
+
+.order-by-select {
+  flex-shrink: 0;
+  width: 100%;
+  height: 42px;
+  border: 1px solid var(--color-white-300);
+  border-radius: 12px;
+  padding: 10px 20px;
+  gap: 10px;
+  font-weight: 400;
+  font-size: 16px;
+  color: var(--color-blue-300);
+
+  @media screen and (max-width: 376px) {
+    display: none;
+  }
+}
+
+.option {
+  font-weight: 400;
+  font-size: 16px;
+  color: var(--color-blue-300);
+
+  @media screen and (max-width: 376px) {
+    font-weight: 400;
+    font-size: 16px;
+    color: var(--color-blue-300);
+  }
+}
+
+.sort-icon {
+  display: none;
+  position: relative;
+  @media screen and (max-width: 376px) {
+    display: block;
+    position: relative;
+    width: 42px;
+    height: 42px;
+    border: 1px solid var(--color-white-300);
+    border-radius: 6px;
+  }
+}
+
+.image-component {
+  position: absolute;
+}
